# Update functional.normalize